Implement services extending from BaseCrudSevice Implement most of the resolvers

+++ backend/service/src/annotations/annotations.resolver.ts
@@ -1,35 +1,82 @@
-import { Resolver, Query, Mutation, Args, Int } from '@nestjs/graphql';
-import { AnnotationsService } from './annotations.service';
 import { Annotation } from './entities/annotation.entity';
+import { AnnotationsService } from './annotations.service';
 import { CreateAnnotationInput } from './dto/create-annotation.input';
+import { Inject } from '@nestjs/common';
+import { PUB_SUB } from 'src/pubsub.module';
+import { RedisPubSub } from 'graphql-redis-subscriptions';
+import { Resolver, Query, Mutation, Args, Int, Subscription } from '@nestjs/graphql';
 import { UpdateAnnotationInput } from './dto/update-annotation.input';
+import { RemoveAnnotationResult } from './dto/remove-annotation-result';
+import { FindAnnotationsArgs } from './dto/find-annotations.args';
+
+const ANNOTATION_ADDED_EVENT = 'annotationAdded';
+const ANNOTATION_UPDATED_EVENT = 'annotationUpdated';
+const ANNOTATION_REMOVED_EVENT = 'annotationRemoved';
 
 @Resolver(() => Annotation)
 export class AnnotationsResolver {
-  constructor(private readonly annotationsService: AnnotationsService) {}
+  constructor(
+    private readonly annotationsService: AnnotationsService,
+    @Inject(PUB_SUB) private readonly pubSub: RedisPubSub,
+  ) {}
+
+  // #region Subscriptions
+  @Subscription(() => Annotation)
+  annotationAdded() {
+    return this.pubSub.asyncIterator(ANNOTATION_ADDED_EVENT);
+  }
+
+  @Subscription(() => Annotation)
+  annotationUpdated() {
+    return this.pubSub.asyncIterator(ANNOTATION_UPDATED_EVENT);
+  }
+
+  @Subscription(() => Annotation)
+  annotationRemoved() {
+    return this.pubSub.asyncIterator(ANNOTATION_REMOVED_EVENT);
+  }
+  // #endregion Subscriptions
+
+  // #region Mutations
+  @Mutation(() => Annotation)
+  async createAnnotation(
+    @Args('createAnnotationInput') createAnnotationInput: CreateAnnotationInput,
+  ): Promise<Annotation> {
+    const annotation = await this.annotationsService.create(createAnnotationInput);
+    annotation && this.pubSub.publish(ANNOTATION_ADDED_EVENT, { [ANNOTATION_ADDED_EVENT]: annotation });
+    return annotation;
+  }
 
   @Mutation(() => Annotation)
-  createAnnotation(@Args('createAnnotationInput') createAnnotationInput: CreateAnnotationInput) {
-    return this.annotationsService.create(createAnnotationInput);
+  async updateAnnotation(
+    @Args('updateAnnotationInput') updateAnnotationInput: UpdateAnnotationInput,
+  ): Promise<Annotation> {
+    const annotation = await this.annotationsService.update(updateAnnotationInput.id, updateAnnotationInput);
+    annotation && this.pubSub.publish(ANNOTATION_UPDATED_EVENT, { [ANNOTATION_UPDATED_EVENT]: annotation });
+    return annotation;
   }
 
-  @Query(() => [Annotation], { name: 'annotations' })
-  findAll() {
-    return this.annotationsService.findAll();
+  @Mutation(() => Annotation)
+  async removeAnnotation(@Args('id', { type: () => Int }) id: number): Promise<RemoveAnnotationResult> {
+    const annotation = await this.annotationsService.findOne(id);
+    if (!annotation) return { result: false };
+
+    const result = await this.annotationsService.remove(id);
+    result && this.pubSub.publish(ANNOTATION_REMOVED_EVENT, { [ANNOTATION_REMOVED_EVENT]: annotation });
+
+    return { result };
+  }
+  // #endregion Mutations
+
+  // #region Queries
+  @Query(() => [Annotation], { name: 'annotations', nullable: true })
+  find(@Args() filter: FindAnnotationsArgs): Promise<Annotation[]> {
+    return this.annotationsService.find(filter);
   }
 
-  @Query(() => Annotation, { name: 'annotation' })
-  findOne(@Args('id', { type: () => Int }) id: number) {
+  @Query(() => Annotation, { name: 'annotation', nullable: true })
+  findOne(@Args('id', { type: () => Int }) id: number): Promise<Annotation> {
     return this.annotationsService.findOne(id);
   }
-
-  @Mutation(() => Annotation)
-  updateAnnotation(@Args('updateAnnotationInput') updateAnnotationInput: UpdateAnnotationInput) {
-    return this.annotationsService.update(updateAnnotationInput.id, updateAnnotationInput);
-  }
-
-  @Mutation(() => Annotation)
-  removeAnnotation(@Args('id', { type: () => Int }) id: number) {
-    return this.annotationsService.remove(id);
-  }
+  // #endregion Queries
 }
