+++ backend/service/src/cease-causes/cease-causes.resolver.ts
@@ -1,35 +1,82 @@
-import { Resolver, Query, Mutation, Args, Int } from '@nestjs/graphql';
+import { Resolver, Query, Mutation, Args, Int, Subscription } from '@nestjs/graphql';
 import { CeaseCausesService } from './cease-causes.service';
 import { CeaseCause } from './entities/cease-cause.entity';
 import { CreateCeaseCauseInput } from './dto/create-cease-cause.input';
 import { UpdateCeaseCauseInput } from './dto/update-cease-cause.input';
+import { Inject } from '@nestjs/common';
+import { PUB_SUB } from 'src/pubsub.module';
+import { RedisPubSub } from 'graphql-redis-subscriptions';
+import { RemoveCeaseCauseResult } from './dto/remove-cease-cuase-result';
+import { FindCeaseCauseArgs } from './dto/find-cease-cause.args';
+
+const CEASE_CAUSE_ADDED_EVENT = 'ceaseCauseAdded';
+const CEASE_CAUSE_UPDATED_EVENT = 'ceaseCauseUpdated';
+const CEASE_CAUSE_REMOVED_EVENT = 'ceaseCauseRemoved';
 
 @Resolver(() => CeaseCause)
 export class CeaseCausesResolver {
-  constructor(private readonly ceaseCausesService: CeaseCausesService) {}
+  constructor(
+    private readonly ceaseCausesService: CeaseCausesService,
+    @Inject(PUB_SUB) private readonly pubSub: RedisPubSub,
+  ) {}
+
+  // #region Subscriptions
+  @Subscription(() => CeaseCause)
+  ceaseCauseAdded() {
+    return this.pubSub.asyncIterator(CEASE_CAUSE_ADDED_EVENT);
+  }
+
+  @Subscription(() => CeaseCause)
+  ceaseCauseUpdated() {
+    return this.pubSub.asyncIterator(CEASE_CAUSE_UPDATED_EVENT);
+  }
+
+  @Subscription(() => CeaseCause)
+  ceaseCauseRemoved() {
+    return this.pubSub.asyncIterator(CEASE_CAUSE_REMOVED_EVENT);
+  }
+  // #endregion Subscriptions
+
+  // #region Mutations
+  @Mutation(() => CeaseCause)
+  async createCeaseCause(
+    @Args('createCeaseCauseInput') createCeaseCauseInput: CreateCeaseCauseInput,
+  ): Promise<CeaseCause> {
+    const ceaseCause = await this.ceaseCausesService.create(createCeaseCauseInput);
+    ceaseCause && this.pubSub.publish(CEASE_CAUSE_ADDED_EVENT, { [CEASE_CAUSE_ADDED_EVENT]: ceaseCause });
+    return ceaseCause;
+  }
 
   @Mutation(() => CeaseCause)
-  createCeaseCause(@Args('createCeaseCauseInput') createCeaseCauseInput: CreateCeaseCauseInput) {
-    return this.ceaseCausesService.create(createCeaseCauseInput);
+  async updateCeaseCause(
+    @Args('updateCeaseCauseInput') updateCeaseCauseInput: UpdateCeaseCauseInput,
+  ): Promise<CeaseCause> {
+    const ceaseCause = await this.ceaseCausesService.update(updateCeaseCauseInput.id, updateCeaseCauseInput);
+    ceaseCause && this.pubSub.publish(CEASE_CAUSE_UPDATED_EVENT, { [CEASE_CAUSE_UPDATED_EVENT]: ceaseCause });
+    return ceaseCause;
   }
 
-  @Query(() => [CeaseCause], { name: 'ceaseCauses' })
-  findAll() {
-    return this.ceaseCausesService.findAll();
+  @Mutation(() => CeaseCause)
+  async removeCeaseCause(@Args('id', { type: () => Int }) id: number): Promise<RemoveCeaseCauseResult> {
+    const ceaseCause = await this.ceaseCausesService.findOne(id);
+    if (!ceaseCause) return { result: false };
+
+    const result = await this.ceaseCausesService.remove(id);
+    result && this.pubSub.publish(CEASE_CAUSE_REMOVED_EVENT, { [CEASE_CAUSE_REMOVED_EVENT]: ceaseCause });
+
+    return { result };
+  }
+  // #endregion Mutations
+
+  // #region Queries
+  @Query(() => [CeaseCause], { name: 'ceaseCauses', nullable: true })
+  find(@Args() filter: FindCeaseCauseArgs): Promise<CeaseCause[]> {
+    return this.ceaseCausesService.find(filter);
   }
 
-  @Query(() => CeaseCause, { name: 'ceaseCause' })
-  findOne(@Args('id', { type: () => Int }) id: number) {
+  @Query(() => CeaseCause, { name: 'ceaseCause', nullable: true })
+  findOne(@Args('id', { type: () => Int }) id: number): Promise<CeaseCause> {
     return this.ceaseCausesService.findOne(id);
   }
-
-  @Mutation(() => CeaseCause)
-  updateCeaseCause(@Args('updateCeaseCauseInput') updateCeaseCauseInput: UpdateCeaseCauseInput) {
-    return this.ceaseCausesService.update(updateCeaseCauseInput.id, updateCeaseCauseInput);
-  }
-
-  @Mutation(() => CeaseCause)
-  removeCeaseCause(@Args('id', { type: () => Int }) id: number) {
-    return this.ceaseCausesService.remove(id);
-  }
+  // #endregion Queries
 }
